Calculating perfect delta_t
This code is much faster. Now, because I know the highest frequency in the signal, I calculate the optimum sampling rate which then determine the sampling interval required to recover the highest frequency in the signal. The code is now updated with this change and it is approximately 30 times faster than what it was before.